paperframe/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pillow==11.1.0
numpy==2.2.2
opencv-python-headless==4.10.0.84
torch==2.2.2
torchvision==0.17.2
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git
transformers==4.47.1
accelerate==1.2.1
